requirements.txt
# Flask app requirements
Flask==2.3.3
Werkzeug==2.3.7

# Original requirements from the whisper-mps project
certifi==2025.1.31
charset-normalizer==3.4.1
decorator==4.4.2
ffmpeg==1.4
filelock==3.17.0
fsspec==2025.2.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.5
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mlx==0.23.1
more-itertools==10.6.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
pillow==11.1.0
proglog==0.1.10
Pygments==2.19.1
pytube==15.0.0
regex==2024.11.6
requests==2.32.3
rich==13.9.4
setuptools==75.8.2
sympy==1.13.1
tiktoken==0.3.3
torch==2.6.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
wheel==0.45.1
whisper-mps @ git+https://github.com/AtomGradient/whisper-mps.git@0c73ed131d0a7e7f16b75e3abdd5a15b77a84b31
torch>=2.0.0
tqdm>=4.64.0
openai-whisper>=20230314
numpy>=1.20.0
ffmpeg-python>=0.2.0
transformers>=4.30.0
setuptools>=65.5.1
wheel>=0.38.0
regex>=2022.0.0